use only double for all coordinates

+++ src/navigation/01_GeoTools.ino
@@ -5,11 +5,11 @@ struct Coord{
   double lng = 0.0;
 };
 
-float get_distance(float lat1, float lng1, float lat2, float lng2){
+float get_distance(double lat1, double lng1, double lat2, double lng2){
   return TinyGPS::distance_between(lat1, lng1, lat2, lng2);
 }
 
-float get_course(float lat1, float lng1, float lat2, float lng2){
+float get_course(double lat1, double lng1, double lat2, double lng2){
   return TinyGPS::course_to(lat1, lng1, lat2, lng2);
 }
